fea: 更换使用Texture作为预览载体，完成滑动预览，删除、拖动时可能有数据问题

+++ app/src/main/java/com/chenfu/avdioedit/Interface/PlayerInterface.java
@@ -20,9 +20,5 @@ public interface PlayerInterface extends Impl {
 
     void pause();
 
-//    void forward10();
-//
-//    void backward10();
-
-    void seekTo(int position);
+    void seekTo(long position);
 }
